+++ src/app/(greeny)/books/CardItem.module.scss
@@ -1,0 +1,61 @@
+@import '@styles/mixin';
+
+@media (max-width: 48rem) {
+  .item_wrapper {
+    @include flex-horizontal(1.8rem);
+    align-items: center;
+
+    .item_image_wrapper {
+      width: 9rem;
+      height: 9rem;
+      img {
+        border-radius: 2rem;
+      }
+    }
+    .item_content {
+      @include flex-vertical(1rem);
+      justify-content: space-between;
+      align-items: flex-start;
+
+      p {
+        font-weight: var(--font-semibold);
+        font-size: 1.4rem;
+      }
+      span {
+        font-weight: var(--font-medium);
+        font-size: 1.2rem;
+      }
+    }
+  }
+}
+
+@media (min-width: 48rem) {
+  li {
+    .item_wrapper {
+      @include flex-vertical(1.6rem);
+      justify-content: space-between;
+      .item_image_wrapper {
+        img {
+          width: 16rem;
+          height: 16rem;
+          border-radius: 2rem;
+        }
+      }
+
+      .item_content {
+        @include flex-vertical(0.8rem);
+        justify-content: space-between;
+        align-items: flex-start;
+
+        p {
+          font-weight: var(--font-semibold);
+          font-size: 1.4rem;
+        }
+        span {
+          font-weight: var(--font-medium);
+          font-size: 1.2rem;
+        }
+      }
+    }
+  }
+}
