+++ src/app/(greeny)/books/CardList.module.scss
@@ -1,0 +1,17 @@
+@import '@styles/mixin';
+
+@media (max-width: 48rem) {
+  .content_list {
+    @include flex-vertical(2rem);
+  }
+}
+
+@media (min-width: 48rem) {
+  .content_list {
+    display: grid;
+    grid-template-columns: repeat(5, 1fr);
+    gap: 2rem;
+
+    //
+  }
+}
